put new tests back in pre-existing modules
Still need to DRY up a few things.
